makes deletion being retried via finalizers by throwing errors on stack state transitions

+++ service/awsconfig/v10/resource/cloudformation/delete.go
@@ -4,10 +4,11 @@ import (
 	"context"
 
 	"github.com/aws/aws-sdk-go/aws"
-	awscloudformation "github.com/aws/aws-sdk-go/service/cloudformation"
-	"github.com/giantswarm/aws-operator/service/awsconfig/v10/key"
+	"github.com/aws/aws-sdk-go/service/cloudformation"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/operatorkit/framework"
+
+	"github.com/giantswarm/aws-operator/service/awsconfig/v10/key"
 )
 
 func (r *Resource) ApplyDeleteChange(ctx context.Context, obj, deleteChange interface{}) error {
@@ -20,10 +21,18 @@ func (r *Resource) ApplyDeleteChange(ctx context.Context, obj, deleteChange inte
 		return microerror.Mask(err)
 	}
 
+	if stackInputToDelete.Status == cloudformation.ResourceStatusUpdateInProgress {
+		r.logger.LogCtx(ctx, "level", "debug", "message", "cannot delete CF stacks due to stack state transitioning")
+
+		// TODO control flow via more proper mechanism via something like the
+		// context like it is done for cancelation already.
+		return microerror.Maskf(deletionMustBeRetriedError, "stack state is transitioning")
+	}
+
 	if stackInputToDelete.Name != "" {
 		r.logger.LogCtx(ctx, "level", "debug", "message", "deleting the guest cluster main stack")
 
-		i := &awscloudformation.DeleteStackInput{
+		i := &cloudformation.DeleteStackInput{
 			StackName: aws.String(key.MainGuestStackName(customObject)),
 		}
 		_, err = r.clients.CloudFormation.DeleteStack(i)
@@ -39,7 +48,7 @@ func (r *Resource) ApplyDeleteChange(ctx context.Context, obj, deleteChange inte
 	if stackInputToDelete.Name != "" {
 		r.logger.LogCtx(ctx, "level", "debug", "message", "deleting the host cluster pre stack")
 
-		i := &awscloudformation.DeleteStackInput{
+		i := &cloudformation.DeleteStackInput{
 			StackName: aws.String(key.MainHostPreStackName(customObject)),
 		}
 		_, err = r.hostClients.CloudFormation.DeleteStack(i)
@@ -55,7 +64,7 @@ func (r *Resource) ApplyDeleteChange(ctx context.Context, obj, deleteChange inte
 	if stackInputToDelete.Name != "" {
 		r.logger.LogCtx(ctx, "level", "debug", "message", "deleting the host cluster post stack")
 
-		i := &awscloudformation.DeleteStackInput{
+		i := &cloudformation.DeleteStackInput{
 			StackName: aws.String(key.MainHostPostStackName(customObject)),
 		}
 		_, err = r.hostClients.CloudFormation.DeleteStack(i)
@@ -84,24 +93,28 @@ func (r *Resource) NewDeletePatch(ctx context.Context, obj, currentState, desire
 }
 
 func (r *Resource) newDeleteChange(ctx context.Context, obj, currentState, desiredState interface{}) (interface{}, error) {
+	currentStackState, err := toStackState(currentState)
+	if err != nil {
+		return StackState{}, microerror.Mask(err)
+	}
 	desiredStackState, err := toStackState(desiredState)
 	if err != nil {
 		return StackState{}, microerror.Mask(err)
 	}
 
-	r.logger.LogCtx(ctx, "level", "debug", "message", "finding out if the guest cluster main stack has to be deleted")
-
-	var deleteState StackState
-
-	if desiredStackState.Name != "" {
-		r.logger.LogCtx(ctx, "level", "debug", "message", "the guest cluster main stack has to be deleted")
-
-		deleteState = StackState{
-			Name: desiredStackState.Name,
-		}
-	} else {
-		r.logger.LogCtx(ctx, "level", "debug", "message", "the guest cluster main stack does not have to be deleted")
+	// For deletion we need the current and desired state in order to process it
+	// reliable. There are cases in which the current state does not contain a
+	// name because of stack state transitions but provides the stack status
+	// itself. The desired state never provides the current stack state in return.
+	// This is why we have to compute the delete state using both entities. Note
+	// that for deletion some properties of the stack state are omitted due to
+	// unimportance to the process.
+	deleteState := StackState{
+		Name:   desiredStackState.Name,
+		Status: currentStackState.Status,
 	}
+
+	r.logger.LogCtx(ctx, "level", "debug", "message", "found the guest cluster main stack that has to be deleted")
 
 	return deleteState, nil
 }
